+++ lib/ftpclient.dart
@@ -2,3 +2,4 @@ library ftpclient;
 
 export 'src/ftpclient_base.dart';
 export 'src/ftpexceptions.dart';
+export 'src/transfermode.dart';